+++ core/model/src/main/java/com/susu/core/model/MyStatistics.kt
@@ -8,10 +8,10 @@ data class MyStatistics(
     val highestAmountSent: StatisticsElement = StatisticsElement(),
     val mostCategory: StatisticsElement = StatisticsElement(),
     val mostRelationship: StatisticsElement = StatisticsElement(),
-    val mostSpentMonth: Int = 0,
+    val mostSpentMonth: Long = 0L,
     val recentSpent: List<StatisticsElement> = emptyList(),
-    val recentTotalSpent: Int = 0,
-    val recentMaximumSpent: Int = 0,
+    val recentTotalSpent: Long = 0L,
+    val recentMaximumSpent: Long = 0L,
 ) {
     companion object {
         val EMPTY = MyStatistics()
